Partial fix for terminal lines becoming obscured
If the height of a putty window was a fraction of the height of the font
being used, up to half of a terminal line could get hidden at the bottom.

This fixes that in a horrifically hacky way, since we don't have easy
access to the height of the font in putty.